backend/requirements.txt
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
colorama==0.4.6
distro==1.9.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
packaging==26.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.12.0
python-dotenv==1.2.2
reportlab==4.4.10
ruff==0.15.5
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
